Refactor App
